+++ src/components/Box/index.ts
@@ -4,4 +4,4 @@ export { default as BoxInfo } from './BoxInfo';
 export { default as BoxInfoSkeleton } from './BoxInfoSkeleton';
 export { default as Comment } from './Comment';
 export { default as Comments } from './Comments';
-export { default as Question } from './Question';
+export { default as QuestionAnswerModal } from './QuestionAnswerModal';
